OrderManagementBackend/requirements.txt
Django==4.2.15
djangorestframework==3.15.2
pillow==10.4.0
pip==22.0.4
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
django-cors-headers==4.4.0
